Refactor: Really separate out selectionUtils to its own file.
